Movement Prototype v.2
Okay guys, in this version:

The blue guy is you. The purple guy is a member of your party. The yellow guy is not one of yours.

If you left-click on a unit, you select the unit. If you right-click a location on the screen, the selected unit will move there.

Things I need to figure out:
1. How to select multiple units via Ctrl+click
2. How to select multiple units via click and drag
3. How to display a graphic that shows you which things you have selected (right now there is no indication)

Feedback is welcome.